requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
bitsandbytes==0.39.1
boto3==1.26.154
botocore==1.29.154
cachetools==5.3.1
charset-normalizer==3.1.0
click==8.0.4
contourpy==1.1.0
cycler==0.11.0
dataclasses==0.6
datasets==2.10.1
dill==0.3.6
distlib==0.3.6
einops==0.4.1
fairscale==0.4.13
faiss-gpu==1.7.2
fastapi==0.85.0
ffmpy==0.3.0
filelock==3.9.0
fire==0.5.0
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.3.0
gensim==4.3.1
git-lfs==1.6
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gradio==3.36.1
gradio_client==0.2.7
grpcio==1.57.0
h11==0.14.0
h5py==3.8.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==2.0.2
lmdb==1.3.0
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
msgpack==1.0.4
multidict==6.0.4
multiprocess==0.70.14
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
orjson==3.9.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
platformdirs==3.1.0
prettytable==3.8.0
protobuf==3.20.3
psutil==5.9.5
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.5
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
pyproject-toml==0.0.11
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-pretrained-bert==0.6.2
pytz==2022.7.1
PyYAML==6.0
ray==2.0.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
s3transfer==0.6.1
safetensors==0.3.1
scikit-learn==1.2.1
scipy==1.10.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
starlette==0.20.4
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6
termcolor==2.3.0
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.10.1+cu111
torchaudio==0.10.1+cu111
torchvision==0.11.2+cu111
tqdm==4.65.0
typing_extensions==4.5.0
uc-micro-py==1.0.2
urllib3==1.26.14
uvicorn==0.18.3
virtualenv==20.20.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.7
wget==3.2
wikiextractor==3.0.6
xxhash==3.2.0
yarl==1.8.2
zipp==3.16.0
